Remove rules for RAML package
I want to remove these rules because that RAML package doesn't have many
installs and it doesn't follow the scope naming guidelines at all. In
case the package will be updated with the correct scopes in the future,
it's unlikely that I would notice it, and then most parts of a file
would be highlighted with wrong colors.