Use distinct colors for --cyanish and --yellowish
Adjusted the colors for --cyanish and --yellowish, which previously were
the same colors as "blue" and "orange". These colors are not used for
syntax highlighting, but can be used by themes and plugins, e.g. for
underlined regions (via the "region.cyanish" and "region.yellowish"
pseudo-scopes) and in minihtml popups. Furthermore they are used to
distinguish kind letters for autocompletions or in the "go to symbol"
panel in ST 4.