Restore color for Git commit hash and undimmed colors in comments
And some other small tweaks.

The commit hash scope was changed in the syntax and the dimmed colors in
comments were accidentally introduced in a previous commit when
generalizing the doc comment scopes.

Alternatively, all colors in Git Commit comments could be dimmed, but
there isn't really a risk of confusion, so I guess the brighter (normal)
colors look better here.