Use base scope instead of language suffix for syntax-specific rules
This ensures that rules still work if further sub-scopes are added to
scope names in syntaxes and sometimes it allows simpler selectors.

Two exceptions where I still kept the language suffix are CSS and
JavaScript, because there are derived syntaxes (TypeScript) or 3rd-party
syntaxes (LESS/Sass) which also use the .js/.css suffix, but a different
base scope.