Adjust some colors for Diff syntax
Before this commit, the from-file and to-file header lines were
highlighted in red and green, similar to how it looks in Adobe Brackets.
I assume the original highlighting is just caused by the + or - signs at
the start of those lines, and not actually intended. I changed it to
purple now, like the filenames in the find-in-files panel, to avoid
confusion with actual removed or added lines.
Also slightly adjusted colors for two other diff scopes.