Small tweak for entity.name scope exclusion
This is a better approach, which allows other "entity.name" scopes to
still be highlighted within comments. Not sure whether there exist any
of that in syntaxes, apart from "comment entity.name.tag" which already
gets its own dimmed color. In case it already exists, or will be added
in the future, it should probably use another rule to apply a dimmed
color for it.