Reduce highlighting for shebang interpreter name
I think the interpreter/language name in shebang comments is not very
important and the highlighting color is more distracting than useful.
Not sure whether it is better to just reduce the opacity of the orange
color, or to remove the color entirely, but for the time being I will
try the latter (compare Markdown fenced code blocks language name).