Add rule for LaTeX formatted math punctuation in Markdown
This aligns the color for the $ symbol which was recently introduced as
LaTeX formatted math delimiter in GitHub Flavored Markdown, with the
color used in LaTeX files. The scope name is from an open pull request
which is expected to be merged in the future, and it is also used in the
MarkdownEditing package.