Remove highlighting for unrecognized types/classes in C#, Go, Java, Ruby
The blue highlighting color should primarily be used for keywords and
some other builtins, but not for arbitrary type/class identifiers. This
also makes highlighting more consistent between languages.

Also a new color variable `type` is added, which will now be used for
unrecognized and user-defined types/classes in these languages, as well
as in JavaScript/TypeScript, PHP and Rust.