Change color of fold marker to make it more noticeable
The color for the fold marker was previously set to grey to match the
style in the original Brackets themes. But the grey color causes it to
not be very easy to recognize within the source code, because it is too
similar to the regular text and comment colors. So it's probably better
to use a distinct color for it, similar to the default color schemes.